backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
sqlalchemy==2.0.23
sqlite3-to-postgres==0.1.2
databases==0.8.0

# ML/AI Dependencies
ultralytics==8.0.220
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
prophet==1.1.5
matplotlib==3.8.2
seaborn==0.13.0
pillow==10.1.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.0
aiohttp==3.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0
